+++ tests/test_cli.py
@@ -17,3 +17,11 @@ def test_scrape_seek():
     with mock.patch("scrape_jobs.sites.seek_com_au.seek_scraper.start") as mock_method:
         scrape_jobs_cli.main([site, config_file])
         mock_method.assert_called_once_with(config_file)
+
+
+def test_scrape_linkedin():
+    site = "linkedin.com"
+    config_file = "scrape.ini"
+    with mock.patch("scrape_jobs.sites.linkedin_com.linkedin_scraper.start") as mock_method:
+        scrape_jobs_cli.main([site, config_file])
+        mock_method.assert_called_once_with(config_file)