+++ CHANGELOG.rst
@@ -13,4 +13,10 @@ Version 1.4.0
 =============
 
 - Restructured common package
-- Implemented pageobjects for linkedin
+- Implemented page objects for linkedin
+
+
+Version 1.5.0
+=============
+
+- support for passing linkedin.com to scrape-jobs CLI